Separate client and server utils
