+++ package.json
@@ -41,7 +41,7 @@
     "axios": "^0.21.2",
     "biggystring": "^3.0.2",
     "body-parser": "^1.19.0",
-    "cleaner-config": "^0.1.9",
+    "cleaner-config": "^0.1.10",
     "cleaners": "^0.3.13",
     "commander": "^6.1.0",
     "cors": "^2.8.5",
